Fixed the data partitioning. Up to now data was only partitioned in a good way if the number of vertices was divisible by the number of available processors. In ALL other cases, the run create one worker PER VERTEX, which is of course highly inefficient.
Fix:
A local constant bool in Graph decides if a worker per single vertex should be created. If that is not the case,
the vertices are distributed as evenly as possible over a number of workers which is equal to the number of processors.
The partitions will generally not have exactly the same size, but sizes will be as close as possible.